requirements.txt
google-genai==1.63.0
yt-dlp==2026.2.4
openai-whisper==20250625
